requirements.txt
# ProofSAR AI — Requirements (Python 3.14 compatible)

# Core Backend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-multipart>=0.0.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.3

# ML & Data Science
scikit-learn>=1.5.0
xgboost>=2.1.0
pandas>=2.2.2
numpy>=2.0.0
joblib>=1.4.0
faker>=24.11.0
imbalanced-learn>=0.12.2
mlflow>=2.13.0


# Explainability
shap>=0.45.1
matplotlib>=3.8.4
seaborn>=0.13.2
plotly>=5.22.0

# Frontend
streamlit>=1.34.0

# PDF Generation
reportlab>=4.2.0

# Audit & Security
cryptography>=42.0.7

# AI / LLM
google-generativeai>=0.5.4

# HTTP
httpx>=0.27.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.1

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.6

# Database & Ops
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
celery>=5.4.0
redis>=5.0.4
psutil>=5.9.8


# Logging
loguru>=0.7.2
